if applied, Moves fileDownload method to a utils folder
