--- app/javascript/dashboard/helper/specs/fileDownload.spec.js
@@ -1,12 +1,0 @@
-/* eslint-disable func-names */
-import downloadFile from '../fileDownload';
-
-describe('#downloadFile', () => {
-  it('should download a file', () => {
-    const file = {
-      name: 'test.txt',
-      content: 'test',
-    };
-    downloadFile(file);
-  });
-});
