+++ app/javascript/dashboard/utils/fileDownload.js
